# use IndexList to improve performance of instance_norm op