--- a/paddle/fluid/operators/instance_norm_op.cc
+++ b/paddle/fluid/operators/instance_norm_op.cc
@@ -180,14 +180,7 @@ class InstanceNormKernel<platform::CPUDeviceContext, T>
 
     auto &dev_ctx = ctx.template device_context<platform::CPUDeviceContext>();
     auto *place = dev_ctx.eigen_device();
-// The IndexList requires a c++11 compliant compiler. If the compiler
-// is older we need to use arrays of indices instead.
-#ifndef EIGEN_HAS_INDEX_LIST
-    Eigen::DSizes<int, 2> bcast(1, sample_size);
-    Eigen::DSizes<int, 2> C_shape(C, 1);
-    Eigen::DSizes<int, 2> NxC_shape(NxC, 1);
-    Eigen::DSizes<int, 1> rdims(1);
-#else
+
     Eigen::IndexList<Eigen::type2index<1>, int> bcast;
     bcast.set(1, sample_size);
     Eigen::IndexList<int, Eigen::type2index<1>> C_shape;
@@ -195,7 +188,7 @@ class InstanceNormKernel<platform::CPUDeviceContext, T>
     Eigen::IndexList<int, Eigen::type2index<1>> NxC_shape;
     NxC_shape.set(0, NxC);
     Eigen::IndexList<Eigen::type2index<1>> rdims;
-#endif
+
     math::SetConstant<platform::CPUDeviceContext, T> set_constant;
 
     saved_mean->mutable_data<T>(ctx.GetPlace());
@@ -328,22 +321,14 @@ class InstanceNormGradKernel<platform::CPUDeviceContext, T>
     Eigen::DSizes<int, 2> rshape(NxC, sample_size);
     Eigen::DSizes<int, 2> param_shape(N, C);
     Eigen::DSizes<int, 2> shape(NxC, sample_size);
-#ifndef EIGEN_HAS_INDEX_LIST
-    Eigen::DSizes<int, 1> rdims(0);
-    Eigen::DSizes<int, 1> mean_rdims(1);
-    Eigen::DSizes<int, 2> bcast(1, sample_size);
-    Eigen::DSizes<int, 2> C_shape(C, 1);
-    Eigen::DSizes<int, 2> NxC_shape(NxC, 1);
-#else
-    Eigen::IndexList<Eigen::type2index<0>, int> rdims;
-    Eigen::IndexList<Eigen::type2index<1>, int> mean_rdims;
+    Eigen::IndexList<Eigen::type2index<0>> rdims;
+    Eigen::IndexList<Eigen::type2index<1>> mean_rdims;
     Eigen::IndexList<Eigen::type2index<1>, int> bcast;
     bcast.set(1, sample_size);
     Eigen::IndexList<int, Eigen::type2index<1>> C_shape;
     C_shape.set(0, C);
     Eigen::IndexList<int, Eigen::type2index<1>> NxC_shape;
     NxC_shape.set(0, NxC);
-#endif
 
     math::SetConstant<platform::CPUDeviceContext, T> set_constant;